Use MongoDB to store rooms
This was very, very hard. I went down a lot of dead ends at first. I don't even remember everything I did to get this working.

I regression tested everything, and everything seems to be in working order.

If anyone's actually reading this commit, you might notice that we're not actually saving the messages yet. This is because they were already wrapped in the room, and it was hard enough getting rooms to work in the first place. I plan on treating messages as a separate entity... soon-ish. Probably next commit.

Also, I renamed the "id" property on the room to "roomId". I was confused, I thought this would get mongo to magically work. It did not - it actually just got worse. (I wish I was using typescript!)